+++ packages/core/src/types/field/field.ts
@@ -1,3 +1,4 @@
+import type { Path } from '../path/index.ts';
 import type { Schema } from '../schema/index.ts';
 import type { Signal } from '../signal/index.ts';
 
@@ -21,6 +22,10 @@ export interface InternalBaseStore {
    * The name of the field.
    */
   name: string;
+  /**
+   * The path to the field.
+   */
+  path: Path;
   /**
    * The schema of the field.
    */